Applied comments
