+++ clients/extension/src/state/utils/getManifestVersion.ts
@@ -1,0 +1,3 @@
+export const getManifestVersion = (): string => {
+  return chrome.runtime.getManifest().version
+}
